refactor: Sections 일급컬렉션 클래스로 변경
